+++ core/gui/src/app/workspace/types/workflow-compiling.interface.ts
